# pkg/api/http/audit/v1/models/audit_instance_d_t_o.go
// Code generated by go-swagger; DO NOT EDIT.

package models

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// AuditInstanceDTO audit instance d t o
//
// swagger:model AuditInstanceDTO
type AuditInstanceDTO struct {

	// instance id
	InstanceID string `json:"instance_id,omitempty"`

	// limit
	Limit int64 `json:"limit,omitempty"`

	// offset
	Offset int64 `json:"offset,omitempty"`

	// type name
	TypeName string `json:"type_name,omitempty"`
}

// Validate validates this audit instance d t o
func (m *AuditInstanceDTO) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this audit instance d t o based on context it is used
func (m *AuditInstanceDTO) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *AuditInstanceDTO) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *AuditInstanceDTO) UnmarshalBinary(b []byte) error {
	var res AuditInstanceDTO
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}