# pkg/api/http/audit/v1/models/audit_period_d_t_o.go
// Code generated by go-swagger; DO NOT EDIT.

package models

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// AuditPeriodDTO audit period d t o
//
// swagger:model AuditPeriodDTO
type AuditPeriodDTO struct {

	// from
	From string `json:"from,omitempty"`

	// limit
	Limit int64 `json:"limit,omitempty"`

	// offset
	Offset int64 `json:"offset,omitempty"`

	// till
	Till string `json:"till,omitempty"`
}

// Validate validates this audit period d t o
func (m *AuditPeriodDTO) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this audit period d t o based on context it is used
func (m *AuditPeriodDTO) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *AuditPeriodDTO) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *AuditPeriodDTO) UnmarshalBinary(b []byte) error {
	var res AuditPeriodDTO
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}