# pkg/api/http/audit/v1/models/audit_user_d_t_o.go
// Code generated by go-swagger; DO NOT EDIT.

package models

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// AuditUserDTO audit user d t o
//
// swagger:model AuditUserDTO
type AuditUserDTO struct {

	// limit
	Limit int64 `json:"limit,omitempty"`

	// offset
	Offset int64 `json:"offset,omitempty"`

	// username
	Username string `json:"username,omitempty"`
}

// Validate validates this audit user d t o
func (m *AuditUserDTO) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this audit user d t o based on context it is used
func (m *AuditUserDTO) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *AuditUserDTO) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *AuditUserDTO) UnmarshalBinary(b []byte) error {
	var res AuditUserDTO
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}