# pkg/api/http/audit/v1/models/data_audit_value_d_t_o.go
// Code generated by go-swagger; DO NOT EDIT.

package models

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// DataAuditValueDTO data audit value d t o
//
// swagger:model DataAuditValueDTO
type DataAuditValueDTO struct {

	// after
	After string `json:"after,omitempty"`

	// before
	Before string `json:"before,omitempty"`

	// description
	Description string `json:"description,omitempty"`

	// name
	Name string `json:"name,omitempty"`
}

// Validate validates this data audit value d t o
func (m *DataAuditValueDTO) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this data audit value d t o based on context it is used
func (m *DataAuditValueDTO) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *DataAuditValueDTO) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *DataAuditValueDTO) UnmarshalBinary(b []byte) error {
	var res DataAuditValueDTO
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}